add roles, security, change controllers, password encryption add UserDetailsServiceImpl

+++ 2.2.2. spring_mvc/src/main/java/web/service/UserService.java
@@ -3,13 +3,14 @@ package web.service;
 import web.model.User;
 
 import java.util.List;
+import java.util.Optional;
 
 public interface UserService {
 
     void addUser(User user);
     void updateUser(User user);
-    void removeUser(int id);
+    void removeUser(Long id);
     List<User> getAllUsers();
-    User getUserById(int id);
+    User getUserById(Long id);
 
 }
