+++ 2.2.2. spring_mvc/src/main/java/web/service/UserServiceImpl.java
@@ -2,13 +2,14 @@ package web.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import web.dao.UserDAO;
 import web.model.User;
 
 import java.util.List;
+
+
 @Service
 public class UserServiceImpl implements UserService {
     @Autowired
@@ -20,24 +21,30 @@ public class UserServiceImpl implements UserService {
     public void addUser(User user) {
         userDAO.addUser(user);
     }
+
     @Transactional
     @Override
     public void updateUser(User user) {
         userDAO.updateUser(user);
     }
+
     @Transactional
     @Override
-    public void removeUser(int id) {
+    public void removeUser(Long id) {
         userDAO.removeUser(id);
     }
+
     @Transactional
     @Override
     public List<User> getAllUsers() {
         return userDAO.getAllUsers();
     }
+
     @Transactional
     @Override
-    public User getUserById(int id) {
+    public User getUserById(Long id) {
         return userDAO.getUserById(id);
     }
+
+
 }
